requirements.txt
datasets>=2.20.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-generativeai>=0.7.0
streamlit>=1.35.0
